# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Usuarios",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
        "/api/orders": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Mostrar todos los pedidos",
                "operationId": "e23a9a38e23c14ee825f28602e7a662b",
                "responses": {
                    "200": {
                        "description": "Retorna todos los pedidos.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "order_number": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "currency": {
                                                "type": "string"
                                            },
                                            "total": {
                                                "type": "number",
                                                "format": "float"
                                            },
                                            "line_items": {
                                                "type": "array",
                                                "items": {
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los pedidos."
                    }
                }
            },
            "post": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Crear un nuevo pedido",
                "operationId": "29a7a02f76402a1d0b5bf5b7f2859efb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "billing",
                                    "shipping",
                                    "line_items",
                                    "total",
                                    "subtotal"
                                ],
                                "properties": {
                                    "billing": {
                                        "type": "object"
                                    },
                                    "shipping": {
                                        "type": "object"
                                    },
                                    "line_items": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        }
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "subtotal": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "discount_total": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "shipping_total": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "payment_method_title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Pedido creado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "order_number": {
                                            "type": "string"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el pedido."
                    }
                }
            }
        },
        "/api/orders/{id}": {
            "get": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Mostrar un pedido específico por ID",
                "operationId": "3380e97d9246b03c46ec4527d9705b30",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Mostrar un pedido específico.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "order_number": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        },
                                        "currency": {
                                            "type": "string"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "float"
                                        },
                                        "line_items": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el pedido."
                    }
                }
            },
            "put": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Actualizar un pedido",
                "operationId": "dde029de3538214545e6b159f9c5a070",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "billing",
                                    "shipping",
                                    "total"
                                ],
                                "properties": {
                                    "billing": {
                                        "type": "object"
                                    },
                                    "shipping": {
                                        "type": "object"
                                    },
                                    "total": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "status": {
                                        "type": "string"
                                    },
                                    "payment_method": {
                                        "type": "string"
                                    },
                                    "payment_method_title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Pedido actualizado correctamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "order_number": {
                                            "type": "string"
                                        },
                                        "total": {
                                            "type": "number",
                                            "format": "float"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el pedido."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Pedidos"
                ],
                "summary": "Eliminar un pedido",
                "operationId": "9783da30b20456f80ff0a022930e2f48",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Pedido eliminado correctamente."
                    },
                    "500": {
                        "description": "Error al eliminar el pedido."
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Pedidos",
            "description": "Pedidos"
        }
    ]
}